+++ srcs/utils.c
@@ -160,10 +160,10 @@ void parse_option(int ac, char **ar){
 
 
 
-void	ft_putnbr_base(uint32_t n, const char *base, uint32_t min_len)
+void	ft_putnbr_base(uint64_t n, const char *base, uint64_t min_len)
 {
   if (min_len != 0){
-    uint32_t n_len = 0, n_copy = n;
+    uint64_t n_len = 0, n_copy = n;
     size_t base_len = strlen(base);
     if (n_copy == 0)
       n_len++;
@@ -171,7 +171,7 @@ void	ft_putnbr_base(uint32_t n, const char *base, uint32_t min_len)
       n_len++;
       n_copy /= base_len;
     }
-    uint32_t i = min_len - n_len;
+    uint64_t i = min_len - n_len;
     while (i-- > 0){
       write(1, "0", 1);
     }
